Start moving configuration things into a subdirectory
