+++ include/configreader.hpp
@@ -7,7 +7,7 @@
 #include <xercesc/parsers/XercesDOMParser.hpp>
 
 #include "xercesguard.hpp"
-#include "xercesstringdeleter.hpp"
+#include "configuration/xercesstringdeleter.hpp"
 #include "controlleditemdata.hpp"
 
 #include "signalheaddata.hpp"
@@ -21,12 +21,12 @@ namespace Signalbox {
   private:
     std::unique_ptr<xercesc::XercesDOMParser> configFileParser;
 
-    std::unique_ptr<XMLCh,xercesstringdeleter> ATTR_id;
+    std::unique_ptr<XMLCh,Configuration::xercesstringdeleter> ATTR_id;
 
-    std::unique_ptr<XMLCh,xercesstringdeleter> ATTR_aspectCount;
+    std::unique_ptr<XMLCh,Configuration::xercesstringdeleter> ATTR_aspectCount;
     
-    std::unique_ptr<XMLCh,xercesstringdeleter> TAG_OutputPin;
-    std::unique_ptr<XMLCh,xercesstringdeleter> ATTR_OutputPin_control;
+    std::unique_ptr<XMLCh,Configuration::xercesstringdeleter> TAG_OutputPin;
+    std::unique_ptr<XMLCh,Configuration::xercesstringdeleter> ATTR_OutputPin_control;
 
     void ReadControlledItems(xercesc::DOMElement* elementSignalbox,
 			     std::vector<std::unique_ptr<ControlledItemData>>& items );
