+++ include/configuration/xercesstringdeleter.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+namespace Signalbox {
+  namespace Configuration {
+    class xercesstringdeleter {
+    public:
+      void operator()( char* str );
+      
+      void operator()( XMLCh* str );
+    };
+  }
+}
