--- include/xercesstringdeleter.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-namespace Signalbox {
-  class xercesstringdeleter {
-  public:
-    void operator()( char* str );
-    
-    void operator()( XMLCh* str );
-  };
-}
